requirements.txt
PyObjC>=9.0
obs-websocket-py>=1.0
PyYAML>=6.0